production_requirements.txt
Django
django-extensions
django-scheduler
django-bower
django-registration
django-guardian
Pillow
requests
celery[redis]
django-rest-framework
gensim
django-guardian
python-ldap
wiki
ipython
gunicorn
django-auth-ldap
psycopg2
git+https://github.com/hltcoe/turkle.git@relocated-admin-models
rdflib
